Refactor regions into partial classes.
