+++ tests/Umbraco.Tests.Integration/Umbraco.Core/Services/ContentPublishingServiceTests.cs
@@ -18,7 +18,7 @@ namespace Umbraco.Cms.Tests.Integration.Umbraco.Core.Services;
     Database = UmbracoTestOptions.Database.NewSchemaPerTest,
     PublishedRepositoryEvents = true,
     WithApplication = true)]
-public class ContentPublishingServiceTests : UmbracoIntegrationTestWithContent
+public partial class ContentPublishingServiceTests : UmbracoIntegrationTestWithContent
 {
     private const string UnknownCulture = "ke-Ke";
 
@@ -38,1303 +38,6 @@ public class ContentPublishingServiceTests : UmbracoIntegrationTestWithContent
     private ITemplateService TemplateService => GetRequiredService<ITemplateService>();
 
     private IContentEditingService ContentEditingService => GetRequiredService<IContentEditingService>();
-
-    #region Publish
-
-    [Test]
-    public async Task Can_Publish_Single_Culture()
-    {
-        var (langEn, langDa, langBe, contentType) = await SetupVariantDoctypeAsync();
-        var setupData = await CreateVariantContentAsync(langEn, langDa, langBe, contentType);
-
-        var publishAttempt = await ContentPublishingService.PublishAsync(
-            setupData.Key,
-            [new() { Culture = langEn.IsoCode }],
-            Constants.Security.SuperUserKey);
-
-        Assert.IsTrue(publishAttempt.Success);
-        var content = ContentService.GetById(setupData.Key);
-        Assert.AreEqual(1, content!.PublishedCultures.Count());
-    }
-
-    [Test]
-    public async Task Can_Publish_Some_Cultures()
-    {
-        var (langEn, langDa, langBe, contentType) = await SetupVariantDoctypeAsync();
-        var setupData = await CreateVariantContentAsync(langEn, langDa, langBe, contentType);
-
-        var publishAttempt = await ContentPublishingService.PublishAsync(
-            setupData.Key,
-            [
-                new() { Culture = langEn.IsoCode }, new() { Culture = langDa.IsoCode },
-            ],
-            Constants.Security.SuperUserKey);
-
-        Assert.IsTrue(publishAttempt.Success);
-        var content = ContentService.GetById(setupData.Key);
-        Assert.AreEqual(2, content!.PublishedCultures.Count());
-    }
-
-    [Test]
-    public async Task Can_Publish_All_Cultures()
-    {
-        var (langEn, langDa, langBe, contentType) = await SetupVariantDoctypeAsync();
-        var setupData = await CreateVariantContentAsync(langEn, langDa, langBe, contentType);
-
-
-        var publishAttempt = await ContentPublishingService.PublishAsync(
-            setupData.Key,
-            [
-                new() { Culture = langEn.IsoCode },
-                new() { Culture = langDa.IsoCode },
-                new() { Culture = langBe.IsoCode },
-            ],
-            Constants.Security.SuperUserKey);
-
-        Assert.IsTrue(publishAttempt.Success);
-        var content = ContentService.GetById(setupData.Key);
-        Assert.AreEqual(3, content!.PublishedCultures.Count());
-    }
-
-    [Test]
-    public async Task Can_NOT_Publish_Invariant_In_Variant_Setup()
-    {
-        var (langEn, langDa, langBe, contentType) = await SetupVariantDoctypeAsync();
-        var setupData = await CreateVariantContentAsync(
-            langEn,
-            langDa,
-            langBe,
-            contentType);
-
-        var publishAttempt = await ContentPublishingService.PublishAsync(
-            setupData.Key,
-            [new() { Culture = Constants.System.InvariantCulture }],
-            Constants.Security.SuperUserKey);
-
-        Assert.IsFalse(publishAttempt.Success);
-        Assert.AreEqual(ContentPublishingOperationStatus.CannotPublishInvariantWhenVariant, publishAttempt.Status);
-
-        var content = ContentService.GetById(setupData.Key);
-        Assert.AreEqual(0, content!.PublishedCultures.Count());
-    }
-
-    [Test]
-    public async Task Can_Publish_Invariant_In_Invariant_Setup()
-    {
-        var doctype = await SetupInvariantDoctypeAsync();
-        var setupData = await CreateInvariantContentAsync(doctype);
-
-        var publishAttempt = await ContentPublishingService.PublishAsync(
-            setupData.Key,
-            [new() { Culture = Constants.System.InvariantCulture }],
-            Constants.Security.SuperUserKey);
-
-        Assert.IsTrue(publishAttempt.Success);
-
-        var content = ContentService.GetById(setupData.Key);
-        Assert.NotNull(content!.PublishDate);
-    }
-    //todo more tests for invariant
-    //todo update schedule date
-
-    [Test]
-    public async Task Can_NOT_Publish_Unknown_Culture()
-    {
-        var (langEn, langDa, langBe, contentType) = await SetupVariantDoctypeAsync();
-        var setupData = await CreateVariantContentAsync(
-            langEn,
-            langDa,
-            langBe,
-            contentType);
-
-        var publishAttempt = await ContentPublishingService.PublishAsync(
-            setupData.Key,
-            [
-                new() { Culture = langEn.IsoCode },
-                new() { Culture = langDa.IsoCode },
-                new() { Culture = UnknownCulture },
-            ],
-            Constants.Security.SuperUserKey);
-
-        Assert.IsFalse(publishAttempt.Success);
-        Assert.AreEqual(ContentPublishingOperationStatus.InvalidCulture, publishAttempt.Status);
-
-        var content = ContentService.GetById(setupData.Key);
-        Assert.AreEqual(0, content!.PublishedCultures.Count());
-    }
-
-    [Test]
-    public async Task Can_NOT_Publish_Scheduled_Culture()
-    {
-        var (langEn, langDa, langBe, contentType) = await SetupVariantDoctypeAsync();
-        var setupData = await CreateVariantContentAsync(
-            langEn,
-            langDa,
-            langBe,
-            contentType);
-
-        var scheduleAttempt = await ContentPublishingService.PublishAsync(
-            setupData.Key,
-            [
-                new()
-                {
-                    Culture = langEn.IsoCode,
-                    Schedule = new ContentScheduleModel { PublishDate = _schedulePublishDate },
-                }
-            ],
-            Constants.Security.SuperUserKey);
-
-        if (scheduleAttempt.Success is false)
-        {
-            throw new Exception("Setup failed");
-        }
-
-        var publishAttempt = await ContentPublishingService.PublishAsync(
-            setupData.Key,
-            [new() { Culture = langEn.IsoCode }],
-            Constants.Security.SuperUserKey);
-
-        Assert.IsFalse(publishAttempt.Success);
-        Assert.AreEqual(ContentPublishingOperationStatus.CultureAwaitingRelease, publishAttempt.Status);
-
-        var content = ContentService.GetById(setupData.Key);
-        Assert.AreEqual(0, content!.PublishedCultures.Count());
-    }
-
-    #endregion
-
-    #region Schedule Publish
-
-    [Test]
-    public async Task Can_Schedule_Publish_Invariant()
-    {
-        var doctype = await SetupInvariantDoctypeAsync();
-        var setupData = await CreateInvariantContentAsync(doctype);
-
-        var scheduleAttempt = await ContentPublishingService.PublishAsync(
-            setupData.Key,
-            [
-                new()
-                {
-                    Culture = Constants.System.InvariantCulture,
-                    Schedule = new ContentScheduleModel { PublishDate = _schedulePublishDate },
-                },
-            ],
-            Constants.Security.SuperUserKey);
-
-        Assert.IsTrue(scheduleAttempt.Success);
-
-        var schedules = ContentService.GetContentScheduleByContentId(setupData.Id);
-        var content = ContentService.GetById(setupData.Key);
-
-        Assert.Multiple(() =>
-        {
-            Assert.AreEqual(0, content!.PublishedCultures.Count());
-            Assert.IsNull(content!.PublishDate);
-            Assert.AreEqual(
-                _schedulePublishDate,
-                schedules.GetSchedule(Constants.System.InvariantCulture, ContentScheduleAction.Release).Single().Date);
-            Assert.AreEqual(1, schedules.FullSchedule.Count);
-        });
-    }
-
-    [Test]
-    public async Task Can_Schedule_Publish_Single_Culture()
-    {
-        var (langEn, langDa, langBe, contentType) = await SetupVariantDoctypeAsync();
-        var setupData = await CreateVariantContentAsync(
-            langEn,
-            langDa,
-            langBe,
-            contentType);
-
-        var scheduleAttempt = await ContentPublishingService.PublishAsync(
-            setupData.Key,
-            [
-                new()
-                {
-                    Culture = langEn.IsoCode,
-                    Schedule = new ContentScheduleModel { PublishDate = _schedulePublishDate },
-                },
-            ],
-            Constants.Security.SuperUserKey);
-
-        Assert.IsTrue(scheduleAttempt.Success);
-
-        var schedules = ContentService.GetContentScheduleByContentId(setupData.Id);
-        var content = ContentService.GetById(setupData.Key);
-
-        Assert.Multiple(() =>
-        {
-            Assert.AreEqual(0, content!.PublishedCultures.Count());
-            Assert.AreEqual(
-                _schedulePublishDate,
-                schedules.GetSchedule(langEn.IsoCode, ContentScheduleAction.Release).Single().Date);
-            Assert.AreEqual(1, schedules.FullSchedule.Count);
-        });
-    }
-
-    [Test]
-    public async Task Can_Schedule_Publish_Some_Cultures()
-    {
-        var (langEn, langDa, langBe, contentType) = await SetupVariantDoctypeAsync();
-        var setupData = await CreateVariantContentAsync(
-            langEn,
-            langDa,
-            langBe,
-            contentType);
-
-        var scheduleAttempt = await ContentPublishingService.PublishAsync(
-            setupData.Key,
-            [
-                new()
-                {
-                    Culture = langEn.IsoCode,
-                    Schedule = new ContentScheduleModel { PublishDate = _schedulePublishDate },
-                },
-                new()
-                {
-                    Culture = langDa.IsoCode,
-                    Schedule = new ContentScheduleModel { PublishDate = _schedulePublishDate },
-                },
-            ],
-            Constants.Security.SuperUserKey);
-
-        Assert.IsTrue(scheduleAttempt.Success);
-
-        var schedules = ContentService.GetContentScheduleByContentId(setupData.Id);
-        var content = ContentService.GetById(setupData.Key);
-
-        Assert.Multiple(() =>
-        {
-            Assert.AreEqual(0, content!.PublishedCultures.Count());
-            Assert.AreEqual(
-                _schedulePublishDate,
-                schedules.GetSchedule(langEn.IsoCode, ContentScheduleAction.Release).Single().Date);
-            Assert.AreEqual(
-                _schedulePublishDate,
-                schedules.GetSchedule(langDa.IsoCode, ContentScheduleAction.Release).Single().Date);
-            Assert.AreEqual(2, schedules.FullSchedule.Count);
-        });
-    }
-
-    [Test]
-    public async Task Can_Schedule_Publish_All_Cultures()
-    {
-        var (langEn, langDa, langBe, contentType) = await SetupVariantDoctypeAsync();
-        var setupData = await CreateVariantContentAsync(
-            langEn,
-            langDa,
-            langBe,
-            contentType);
-
-        var scheduleAttempt = await ContentPublishingService.PublishAsync(
-            setupData.Key,
-            [
-                new()
-                {
-                    Culture = langEn.IsoCode,
-                    Schedule = new ContentScheduleModel { PublishDate = _schedulePublishDate },
-                },
-                new()
-                {
-                    Culture = langDa.IsoCode,
-                    Schedule = new ContentScheduleModel { PublishDate = _schedulePublishDate },
-                },
-                new()
-                {
-                    Culture = langBe.IsoCode,
-                    Schedule = new ContentScheduleModel { PublishDate = _schedulePublishDate },
-                },
-            ],
-            Constants.Security.SuperUserKey);
-
-        Assert.IsTrue(scheduleAttempt.Success);
-
-        var schedules = ContentService.GetContentScheduleByContentId(setupData.Id);
-        var content = ContentService.GetById(setupData.Key);
-
-        Assert.Multiple(() =>
-        {
-            Assert.AreEqual(0, content!.PublishedCultures.Count());
-            Assert.AreEqual(
-                _schedulePublishDate,
-                schedules.GetSchedule(langEn.IsoCode, ContentScheduleAction.Release).Single().Date);
-            Assert.AreEqual(
-                _schedulePublishDate,
-                schedules.GetSchedule(langDa.IsoCode, ContentScheduleAction.Release).Single().Date);
-            Assert.AreEqual(
-                _schedulePublishDate,
-                schedules.GetSchedule(langBe.IsoCode, ContentScheduleAction.Release).Single().Date);
-            Assert.AreEqual(3, schedules.FullSchedule.Count);
-        });
-    }
-
-    [Test]
-    public async Task Can_NOT_Schedule_Publish_Unknown_Culture()
-    {
-        var (langEn, langDa, langBe, contentType) = await SetupVariantDoctypeAsync();
-        var setupData = await CreateVariantContentAsync(
-            langEn,
-            langDa,
-            langBe,
-            contentType);
-
-        var scheduleAttempt = await ContentPublishingService.PublishAsync(
-            setupData.Key,
-            [
-                new()
-                {
-                    Culture = langEn.IsoCode,
-                    Schedule = new ContentScheduleModel { PublishDate = _schedulePublishDate },
-                },
-                new()
-                {
-                    Culture = langDa.IsoCode,
-                    Schedule = new ContentScheduleModel { PublishDate = _schedulePublishDate },
-                },
-                new()
-                {
-                    Culture = UnknownCulture,
-                    Schedule = new ContentScheduleModel { PublishDate = _schedulePublishDate }
-                },
-            ],
-            Constants.Security.SuperUserKey);
-
-        Assert.IsFalse(scheduleAttempt.Success);
-        Assert.AreEqual(ContentPublishingOperationStatus.InvalidCulture, scheduleAttempt.Status);
-
-        var schedules = ContentService.GetContentScheduleByContentId(setupData.Id);
-        var content = ContentService.GetById(setupData.Key);
-
-        Assert.Multiple(() =>
-        {
-            Assert.AreEqual(0, content!.PublishedCultures.Count());
-            Assert.AreEqual(0, schedules.FullSchedule.Count);
-        });
-    }
-
-    #endregion
-
-    #region Schedule Unpublish
-
-    [Test]
-    public async Task Can_Schedule_Unpublish_Invariant()
-    {
-        var doctype = await SetupInvariantDoctypeAsync();
-        var setupData = await CreateInvariantContentAsync(doctype);
-
-        var scheduleAttempt = await ContentPublishingService.PublishAsync(
-            setupData.Key,
-            [
-                new()
-                {
-                    Culture = Constants.System.InvariantCulture,
-                    Schedule = new ContentScheduleModel { UnpublishDate = _scheduleUnPublishDate },
-                },
-            ],
-            Constants.Security.SuperUserKey);
-
-        Assert.IsTrue(scheduleAttempt.Success);
-
-        var schedules = ContentService.GetContentScheduleByContentId(setupData.Id);
-        var content = ContentService.GetById(setupData.Key);
-
-        Assert.Multiple(() =>
-        {
-            Assert.AreEqual(0, content!.PublishedCultures.Count());
-            Assert.IsNull(content!.PublishDate);
-            Assert.AreEqual(
-                _scheduleUnPublishDate,
-                schedules.GetSchedule(Constants.System.InvariantCulture, ContentScheduleAction.Expire).Single().Date);
-            Assert.AreEqual(1, schedules.FullSchedule.Count);
-        });
-    }
-
-    [Test]
-    public async Task Can_Schedule_Unpublish_Single_Culture()
-    {
-        var (langEn, langDa, langBe, contentType) = await SetupVariantDoctypeAsync();
-        var setupData = await CreateVariantContentAsync(
-            langEn,
-            langDa,
-            langBe,
-            contentType);
-
-        var scheduleAttempt = await ContentPublishingService.PublishAsync(
-            setupData.Key,
-            [
-                new()
-                {
-                    Culture = langEn.IsoCode,
-                    Schedule = new ContentScheduleModel { UnpublishDate = _schedulePublishDate },
-                },
-            ],
-            Constants.Security.SuperUserKey);
-
-        Assert.IsTrue(scheduleAttempt.Success);
-
-        var schedules = ContentService.GetContentScheduleByContentId(setupData.Id);
-        var content = ContentService.GetById(setupData.Key);
-
-        Assert.Multiple(() =>
-        {
-            Assert.AreEqual(0, content!.PublishedCultures.Count());
-            Assert.AreEqual(
-                _schedulePublishDate,
-                schedules.GetSchedule(langEn.IsoCode, ContentScheduleAction.Expire).Single().Date);
-            Assert.AreEqual(1, schedules.FullSchedule.Count);
-        });
-    }
-
-    [Test]
-    public async Task Can_Schedule_Unpublish_Some_Cultures()
-    {
-        var (langEn, langDa, langBe, contentType) = await SetupVariantDoctypeAsync();
-        var setupData = await CreateVariantContentAsync(
-            langEn,
-            langDa,
-            langBe,
-            contentType);
-
-        var scheduleAttempt = await ContentPublishingService.PublishAsync(
-            setupData.Key,
-            [
-                new()
-                {
-                    Culture = langEn.IsoCode,
-                    Schedule = new ContentScheduleModel { UnpublishDate = _schedulePublishDate },
-                },
-                new()
-                {
-                    Culture = langDa.IsoCode,
-                    Schedule = new ContentScheduleModel { UnpublishDate = _schedulePublishDate },
-                },
-            ],
-            Constants.Security.SuperUserKey);
-
-        Assert.IsTrue(scheduleAttempt.Success);
-
-        var schedules = ContentService.GetContentScheduleByContentId(setupData.Id);
-        var content = ContentService.GetById(setupData.Key);
-
-        Assert.Multiple(() =>
-        {
-            Assert.AreEqual(0, content!.PublishedCultures.Count());
-            Assert.AreEqual(
-                _schedulePublishDate,
-                schedules.GetSchedule(langEn.IsoCode, ContentScheduleAction.Expire).Single().Date);
-            Assert.AreEqual(
-                _schedulePublishDate,
-                schedules.GetSchedule(langDa.IsoCode, ContentScheduleAction.Expire).Single().Date);
-            Assert.AreEqual(2, schedules.FullSchedule.Count);
-        });
-    }
-
-    [Test]
-    public async Task Can_Schedule_Unpublish_All_Cultures()
-    {
-        var (langEn, langDa, langBe, contentType) = await SetupVariantDoctypeAsync();
-        var setupData = await CreateVariantContentAsync(
-            langEn,
-            langDa,
-            langBe,
-            contentType);
-
-        var scheduleAttempt = await ContentPublishingService.PublishAsync(
-            setupData.Key,
-            [
-                new()
-                {
-                    Culture = langEn.IsoCode,
-                    Schedule = new ContentScheduleModel { UnpublishDate = _schedulePublishDate },
-                },
-                new()
-                {
-                    Culture = langDa.IsoCode,
-                    Schedule = new ContentScheduleModel { UnpublishDate = _schedulePublishDate },
-                },
-                new()
-                {
-                    Culture = langBe.IsoCode,
-                    Schedule = new ContentScheduleModel { UnpublishDate = _schedulePublishDate },
-                },
-            ],
-            Constants.Security.SuperUserKey);
-
-        Assert.IsTrue(scheduleAttempt.Success);
-
-        var schedules = ContentService.GetContentScheduleByContentId(setupData.Id);
-        var content = ContentService.GetById(setupData.Key);
-
-        Assert.Multiple(() =>
-        {
-            Assert.AreEqual(0, content!.PublishedCultures.Count());
-            Assert.AreEqual(
-                _schedulePublishDate,
-                schedules.GetSchedule(langEn.IsoCode, ContentScheduleAction.Expire).Single().Date);
-            Assert.AreEqual(
-                _schedulePublishDate,
-                schedules.GetSchedule(langDa.IsoCode, ContentScheduleAction.Expire).Single().Date);
-            Assert.AreEqual(
-                _schedulePublishDate,
-                schedules.GetSchedule(langBe.IsoCode, ContentScheduleAction.Expire).Single().Date);
-            Assert.AreEqual(3, schedules.FullSchedule.Count);
-        });
-    }
-
-    [Test]
-    public async Task Can_NOT_Schedule_Unpublish_Unknown_Culture()
-    {
-        var (langEn, langDa, langBe, contentType) = await SetupVariantDoctypeAsync();
-        var setupData = await CreateVariantContentAsync(
-            langEn,
-            langDa,
-            langBe,
-            contentType);
-
-        var scheduleAttempt = await ContentPublishingService.PublishAsync(
-            setupData.Key,
-            [
-                new()
-                {
-                    Culture = langEn.IsoCode,
-                    Schedule = new ContentScheduleModel { UnpublishDate = _schedulePublishDate },
-                },
-                new()
-                {
-                    Culture = langDa.IsoCode,
-                    Schedule = new ContentScheduleModel { UnpublishDate = _schedulePublishDate },
-                },
-                new()
-                {
-                    Culture = UnknownCulture,
-                    Schedule = new ContentScheduleModel { UnpublishDate = _schedulePublishDate }
-                },
-            ],
-            Constants.Security.SuperUserKey);
-
-        Assert.IsFalse(scheduleAttempt.Success);
-        Assert.AreEqual(ContentPublishingOperationStatus.InvalidCulture, scheduleAttempt.Status);
-
-        var schedules = ContentService.GetContentScheduleByContentId(setupData.Id);
-        var content = ContentService.GetById(setupData.Key);
-
-        Assert.Multiple(() =>
-        {
-            Assert.AreEqual(0, content!.PublishedCultures.Count());
-            Assert.AreEqual(0, schedules.FullSchedule.Count);
-        });
-    }
-
-    #endregion
-
-    #region Unschedule Publish
-
-    [Test]
-    public async Task Can_UnSchedule_Publish_Invariant()
-    {
-        var doctype = await SetupInvariantDoctypeAsync();
-        var setupData = await CreateInvariantContentAsync(doctype);
-
-        var scheduleSetupAttempt =
-            await SchedulePublishAndUnPublishInvariantAsync(setupData);
-
-        if (scheduleSetupAttempt.Success is false)
-        {
-            throw new Exception("Setup failed");
-        }
-
-        var unscheduleAttempt = await ContentPublishingService.PublishAsync(
-            setupData.Key,
-            [
-                new()
-                {
-                    Culture = Constants.System.InvariantCulture,
-                    Schedule = new ContentScheduleModel { UnpublishDate = _scheduleUnPublishDate },
-                },
-            ],
-            Constants.Security.SuperUserKey);
-
-        Assert.IsTrue(unscheduleAttempt.Success);
-
-        var schedules = ContentService.GetContentScheduleByContentId(setupData.Id);
-        var content = ContentService.GetById(setupData.Key);
-
-        Assert.Multiple(() =>
-        {
-            Assert.AreEqual(0, content!.PublishedCultures.Count());
-            Assert.IsNull(content!.PublishDate);
-            Assert.IsFalse(schedules.GetSchedule(Constants.System.InvariantCulture, ContentScheduleAction.Release).Any());
-            Assert.IsTrue(schedules.GetSchedule(Constants.System.InvariantCulture, ContentScheduleAction.Expire).Any());
-            Assert.AreEqual(1, schedules.FullSchedule.Count);
-        });
-    }
-
-    [Test]
-    public async Task Can_Unschedule_Publish_Single_Culture()
-    {
-        var setupInfo = await SetupVariantDoctypeAsync();
-        var setupData = await CreateVariantContentAsync(
-            setupInfo.LangEn,
-            setupInfo.LangDa,
-            setupInfo.LangBe,
-            setupInfo.contentType);
-
-        var scheduleSetupAttempt =
-            await SchedulePublishAndUnPublishForAllCulturesAsync(setupData, setupInfo);
-
-        if (scheduleSetupAttempt.Success is false)
-        {
-            throw new Exception("Setup failed");
-        }
-
-        var scheduleAttempt = await ContentPublishingService.PublishAsync(
-            setupData.Key,
-            [
-                new()
-                {
-                    Culture = setupInfo.LangEn.IsoCode,
-                    Schedule = new ContentScheduleModel { UnpublishDate = _scheduleUnPublishDate },
-                },
-            ],
-            Constants.Security.SuperUserKey);
-
-        Assert.IsTrue(scheduleAttempt.Success);
-
-        var schedules = ContentService.GetContentScheduleByContentId(setupData.Id);
-        var content = ContentService.GetById(setupData.Key);
-
-        Assert.Multiple(() =>
-        {
-            Assert.AreEqual(0, content!.PublishedCultures.Count());
-            Assert.IsFalse(schedules.GetSchedule(setupInfo.LangEn.IsoCode, ContentScheduleAction.Release).Any());
-            Assert.AreEqual(5, schedules.FullSchedule.Count);
-        });
-    }
-
-    [Test]
-    public async Task Can_Unschedule_Publish_Some_Cultures()
-    {
-        var setupInfo = await SetupVariantDoctypeAsync();
-        var setupData = await CreateVariantContentAsync(
-            setupInfo.LangEn,
-            setupInfo.LangDa,
-            setupInfo.LangBe,
-            setupInfo.contentType);
-
-        var scheduleSetupAttempt =
-            await SchedulePublishAndUnPublishForAllCulturesAsync(setupData, setupInfo);
-
-        if (scheduleSetupAttempt.Success is false)
-        {
-            throw new Exception("Setup failed");
-        }
-
-        var scheduleAttempt = await ContentPublishingService.PublishAsync(
-            setupData.Key,
-            [
-                new()
-                {
-                    Culture = setupInfo.LangEn.IsoCode,
-                    Schedule = new ContentScheduleModel { UnpublishDate = _scheduleUnPublishDate }
-                },
-                new()
-                {
-                    Culture = setupInfo.LangDa.IsoCode,
-                    Schedule = new ContentScheduleModel { UnpublishDate = _scheduleUnPublishDate }
-                },
-            ],
-            Constants.Security.SuperUserKey);
-
-        Assert.IsTrue(scheduleAttempt.Success);
-
-        var schedules = ContentService.GetContentScheduleByContentId(setupData.Id);
-        var content = ContentService.GetById(setupData.Key);
-
-        Assert.Multiple(() =>
-        {
-            Assert.AreEqual(0, content!.PublishedCultures.Count());
-            Assert.IsFalse(schedules.GetSchedule(setupInfo.LangEn.IsoCode, ContentScheduleAction.Release).Any());
-            Assert.IsFalse(schedules.GetSchedule(setupInfo.LangDa.IsoCode, ContentScheduleAction.Release).Any());
-            Assert.AreEqual(4, schedules.FullSchedule.Count);
-        });
-    }
-
-    [Test]
-    public async Task Can_Unschedule_Publish_All_Cultures()
-    {
-        var setupInfo = await SetupVariantDoctypeAsync();
-        var setupData = await CreateVariantContentAsync(
-            setupInfo.LangEn,
-            setupInfo.LangDa,
-            setupInfo.LangBe,
-            setupInfo.contentType);
-
-        var scheduleSetupAttempt =
-            await SchedulePublishAndUnPublishForAllCulturesAsync(setupData, setupInfo);
-
-        if (scheduleSetupAttempt.Success is false)
-        {
-            throw new Exception("Setup failed");
-        }
-
-        var scheduleAttempt = await ContentPublishingService.PublishAsync(
-            setupData.Key,
-            [
-                new()
-                {
-                    Culture = setupInfo.LangEn.IsoCode,
-                    Schedule = new ContentScheduleModel { UnpublishDate = _scheduleUnPublishDate },
-                },
-                new()
-                {
-                    Culture = setupInfo.LangDa.IsoCode,
-                    Schedule = new ContentScheduleModel { UnpublishDate = _scheduleUnPublishDate },
-                },
-                new()
-                {
-                    Culture = setupInfo.LangBe.IsoCode,
-                    Schedule = new ContentScheduleModel { UnpublishDate = _scheduleUnPublishDate },
-                },
-            ],
-            Constants.Security.SuperUserKey);
-
-        Assert.IsTrue(scheduleAttempt.Success);
-
-        var schedules = ContentService.GetContentScheduleByContentId(setupData.Id);
-        var content = ContentService.GetById(setupData.Key);
-
-        Assert.Multiple(() =>
-        {
-            Assert.AreEqual(0, content!.PublishedCultures.Count());
-            Assert.IsFalse(schedules.GetSchedule(setupInfo.LangEn.IsoCode, ContentScheduleAction.Release).Any());
-            Assert.IsFalse(schedules.GetSchedule(setupInfo.LangDa.IsoCode, ContentScheduleAction.Release).Any());
-            Assert.IsFalse(schedules.GetSchedule(setupInfo.LangBe.IsoCode, ContentScheduleAction.Release).Any());
-            Assert.AreEqual(3, schedules.FullSchedule.Count);
-        });
-    }
-
-    [Test]
-    public async Task Can_NOT_Unschedule_Publish_Unknown_Culture()
-    {
-        var setupInfo = await SetupVariantDoctypeAsync();
-        var setupData = await CreateVariantContentAsync(
-            setupInfo.LangEn,
-            setupInfo.LangDa,
-            setupInfo.LangBe,
-            setupInfo.contentType);
-
-        var scheduleSetupAttempt =
-            await SchedulePublishAndUnPublishForAllCulturesAsync(setupData, setupInfo);
-
-        if (scheduleSetupAttempt.Success is false)
-        {
-            throw new Exception("Setup failed");
-        }
-
-        var scheduleAttempt = await ContentPublishingService.PublishAsync(
-            setupData.Key,
-            [
-                new()
-                {
-                    Culture = setupInfo.LangEn.IsoCode,
-                    Schedule = new ContentScheduleModel { UnpublishDate = _scheduleUnPublishDate },
-                },
-                new()
-                {
-                    Culture = setupInfo.LangDa.IsoCode,
-                    Schedule = new ContentScheduleModel { UnpublishDate = _scheduleUnPublishDate },
-                },
-                new()
-                {
-                    Culture = UnknownCulture,
-                    Schedule = new ContentScheduleModel { UnpublishDate = _scheduleUnPublishDate },
-                },
-            ],
-            Constants.Security.SuperUserKey);
-
-        Assert.IsFalse(scheduleAttempt.Success);
-        Assert.AreEqual(ContentPublishingOperationStatus.InvalidCulture, scheduleAttempt.Status);
-
-        var schedules = ContentService.GetContentScheduleByContentId(setupData.Id);
-        var content = ContentService.GetById(setupData.Key);
-
-        Assert.Multiple(() =>
-        {
-            Assert.AreEqual(0, content!.PublishedCultures.Count());
-            Assert.AreEqual(6, schedules.FullSchedule.Count);
-        });
-    }
-
-    #endregion
-
-    #region Unschedule Unpublish
-
-    [Test]
-    public async Task Can_UnSchedule_Unpublish_Invariant()
-    {
-        var doctype = await SetupInvariantDoctypeAsync();
-        var setupData = await CreateInvariantContentAsync(doctype);
-
-        var scheduleSetupAttempt =
-            await SchedulePublishAndUnPublishInvariantAsync(setupData);
-
-        if (scheduleSetupAttempt.Success is false)
-        {
-            throw new Exception("Setup failed");
-        }
-
-        var unscheduleAttempt = await ContentPublishingService.PublishAsync(
-            setupData.Key,
-            [
-                new()
-                {
-                    Culture = Constants.System.InvariantCulture,
-                    Schedule = new ContentScheduleModel { PublishDate = _schedulePublishDate },
-                },
-            ],
-            Constants.Security.SuperUserKey);
-
-        Assert.IsTrue(unscheduleAttempt.Success);
-
-        var schedules = ContentService.GetContentScheduleByContentId(setupData.Id);
-        var content = ContentService.GetById(setupData.Key);
-
-        Assert.Multiple(() =>
-        {
-            Assert.AreEqual(0, content!.PublishedCultures.Count());
-            Assert.IsNull(content!.PublishDate);
-            Assert.IsFalse(schedules.GetSchedule(Constants.System.InvariantCulture, ContentScheduleAction.Expire).Any());
-            Assert.IsTrue(schedules.GetSchedule(Constants.System.InvariantCulture, ContentScheduleAction.Release).Any());
-            Assert.AreEqual(1, schedules.FullSchedule.Count);
-        });
-    }
-
-    [Test]
-    public async Task Can_Unschedule_Unpublish_Single_Culture()
-    {
-        var setupInfo = await SetupVariantDoctypeAsync();
-        var setupData = await CreateVariantContentAsync(
-            setupInfo.LangEn,
-            setupInfo.LangDa,
-            setupInfo.LangBe,
-            setupInfo.contentType);
-
-        var scheduleSetupAttempt =
-            await SchedulePublishAndUnPublishForAllCulturesAsync(setupData, setupInfo);
-
-        if (scheduleSetupAttempt.Success is false)
-        {
-            throw new Exception("Setup failed");
-        }
-
-        var scheduleAttempt = await ContentPublishingService.PublishAsync(
-            setupData.Key,
-            [
-                new()
-                {
-                    Culture = setupInfo.LangEn.IsoCode,
-                    Schedule = new ContentScheduleModel { PublishDate = _schedulePublishDate },
-                },
-            ],
-            Constants.Security.SuperUserKey);
-
-        Assert.IsTrue(scheduleAttempt.Success);
-
-        var schedules = ContentService.GetContentScheduleByContentId(setupData.Id);
-        var content = ContentService.GetById(setupData.Key);
-
-        Assert.Multiple(() =>
-        {
-            Assert.AreEqual(0, content!.PublishedCultures.Count());
-            Assert.IsFalse(schedules.GetSchedule(setupInfo.LangEn.IsoCode, ContentScheduleAction.Expire).Any());
-            Assert.AreEqual(5, schedules.FullSchedule.Count);
-        });
-    }
-
-    [Test]
-    public async Task Can_Unschedule_Unpublish_Some_Cultures()
-    {
-        var setupInfo = await SetupVariantDoctypeAsync();
-        var setupData = await CreateVariantContentAsync(
-            setupInfo.LangEn,
-            setupInfo.LangDa,
-            setupInfo.LangBe,
-            setupInfo.contentType);
-
-        var scheduleSetupAttempt =
-            await SchedulePublishAndUnPublishForAllCulturesAsync(setupData, setupInfo);
-
-        if (scheduleSetupAttempt.Success is false)
-        {
-            throw new Exception("Setup failed");
-        }
-
-        var scheduleAttempt = await ContentPublishingService.PublishAsync(
-            setupData.Key,
-            [
-                new()
-                {
-                    Culture = setupInfo.LangEn.IsoCode,
-                    Schedule = new ContentScheduleModel { PublishDate = _schedulePublishDate }
-                },
-                new()
-                {
-                    Culture = setupInfo.LangDa.IsoCode,
-                    Schedule = new ContentScheduleModel { PublishDate = _schedulePublishDate }
-                },
-            ],
-            Constants.Security.SuperUserKey);
-
-        Assert.IsTrue(scheduleAttempt.Success);
-
-        var schedules = ContentService.GetContentScheduleByContentId(setupData.Id);
-        var content = ContentService.GetById(setupData.Key);
-
-        Assert.Multiple(() =>
-        {
-            Assert.AreEqual(0, content!.PublishedCultures.Count());
-            Assert.IsFalse(schedules.GetSchedule(setupInfo.LangEn.IsoCode, ContentScheduleAction.Expire).Any());
-            Assert.IsFalse(schedules.GetSchedule(setupInfo.LangDa.IsoCode, ContentScheduleAction.Expire).Any());
-            Assert.AreEqual(4, schedules.FullSchedule.Count);
-        });
-    }
-
-    [Test]
-    public async Task Can_Unschedule_Unpublish_All_Cultures()
-    {
-        var setupInfo = await SetupVariantDoctypeAsync();
-        var setupData = await CreateVariantContentAsync(
-            setupInfo.LangEn,
-            setupInfo.LangDa,
-            setupInfo.LangBe,
-            setupInfo.contentType);
-
-        var scheduleSetupAttempt =
-            await SchedulePublishAndUnPublishForAllCulturesAsync(setupData, setupInfo);
-
-        if (scheduleSetupAttempt.Success is false)
-        {
-            throw new Exception("Setup failed");
-        }
-
-        var scheduleAttempt = await ContentPublishingService.PublishAsync(
-            setupData.Key,
-            [
-                new()
-                {
-                    Culture = setupInfo.LangEn.IsoCode,
-                    Schedule = new ContentScheduleModel { PublishDate = _schedulePublishDate },
-                },
-                new()
-                {
-                    Culture = setupInfo.LangDa.IsoCode,
-                    Schedule = new ContentScheduleModel { PublishDate = _schedulePublishDate },
-                },
-                new()
-                {
-                    Culture = setupInfo.LangBe.IsoCode,
-                    Schedule = new ContentScheduleModel { PublishDate = _schedulePublishDate },
-                },
-            ],
-            Constants.Security.SuperUserKey);
-
-        Assert.IsTrue(scheduleAttempt.Success);
-
-        var schedules = ContentService.GetContentScheduleByContentId(setupData.Id);
-        var content = ContentService.GetById(setupData.Key);
-
-        Assert.Multiple(() =>
-        {
-            Assert.AreEqual(0, content!.PublishedCultures.Count());
-            Assert.IsFalse(schedules.GetSchedule(setupInfo.LangEn.IsoCode, ContentScheduleAction.Expire).Any());
-            Assert.IsFalse(schedules.GetSchedule(setupInfo.LangDa.IsoCode, ContentScheduleAction.Expire).Any());
-            Assert.IsFalse(schedules.GetSchedule(setupInfo.LangBe.IsoCode, ContentScheduleAction.Expire).Any());
-            Assert.AreEqual(3, schedules.FullSchedule.Count);
-        });
-    }
-
-    [Test]
-    public async Task Can_NOT_Unschedule_Unpublish_Unknown_Culture()
-    {
-        var setupInfo = await SetupVariantDoctypeAsync();
-        var setupData = await CreateVariantContentAsync(
-            setupInfo.LangEn,
-            setupInfo.LangDa,
-            setupInfo.LangBe,
-            setupInfo.contentType);
-
-        var scheduleSetupAttempt =
-            await SchedulePublishAndUnPublishForAllCulturesAsync(setupData, setupInfo);
-
-        if (scheduleSetupAttempt.Success is false)
-        {
-            throw new Exception("Setup failed");
-        }
-
-        var scheduleAttempt = await ContentPublishingService.PublishAsync(
-            setupData.Key,
-            [
-                new()
-                {
-                    Culture = setupInfo.LangEn.IsoCode,
-                    Schedule = new ContentScheduleModel { PublishDate = _schedulePublishDate },
-                },
-                new()
-                {
-                    Culture = setupInfo.LangDa.IsoCode,
-                    Schedule = new ContentScheduleModel { PublishDate = _schedulePublishDate },
-                },
-                new()
-                {
-                    Culture = UnknownCulture,
-                    Schedule = new ContentScheduleModel { PublishDate = _schedulePublishDate },
-                },
-            ],
-            Constants.Security.SuperUserKey);
-
-        Assert.IsFalse(scheduleAttempt.Success);
-        Assert.AreEqual(ContentPublishingOperationStatus.InvalidCulture, scheduleAttempt.Status);
-        var schedules = ContentService.GetContentScheduleByContentId(setupData.Id);
-        var content = ContentService.GetById(setupData.Key);
-
-        Assert.Multiple(() =>
-        {
-            Assert.AreEqual(0, content!.PublishedCultures.Count());
-            Assert.AreEqual(6, schedules.FullSchedule.Count);
-        });
-    }
-
-    #endregion
-
-    #region Clean Schedule
-
-    [Test]
-    public async Task Can_Clear_Schedule_Invariant()
-    {
-        var doctype = await SetupInvariantDoctypeAsync();
-        var setupData = await CreateInvariantContentAsync(doctype);
-
-        var scheduleSetupAttempt =
-            await SchedulePublishAndUnPublishInvariantAsync(setupData);
-
-        if (scheduleSetupAttempt.Success is false)
-        {
-            throw new Exception("Setup failed");
-        }
-
-        var clearScheduleAttempt = await ContentPublishingService.PublishAsync(
-            setupData.Key,
-            [
-                new()
-                {
-                    Culture = Constants.System.InvariantCulture,
-                    Schedule = new ContentScheduleModel(),
-                },
-            ],
-            Constants.Security.SuperUserKey);
-
-        Assert.IsTrue(clearScheduleAttempt.Success);
-
-        var schedules = ContentService.GetContentScheduleByContentId(setupData.Id);
-        var content = ContentService.GetById(setupData.Key);
-
-        Assert.Multiple(() =>
-        {
-            Assert.AreEqual(0, content!.PublishedCultures.Count());
-            Assert.IsNull(content!.PublishDate);
-            Assert.AreEqual(0, schedules.FullSchedule.Count);
-        });
-    }
-
-    [Test]
-    public async Task Can_Clear_Schedule_Single_Culture()
-    {
-        var setupInfo = await SetupVariantDoctypeAsync();
-        var setupData = await CreateVariantContentAsync(
-            setupInfo.LangEn,
-            setupInfo.LangDa,
-            setupInfo.LangBe,
-            setupInfo.contentType);
-
-        var scheduleSetupAttempt =
-            await SchedulePublishAndUnPublishForAllCulturesAsync(setupData, setupInfo);
-
-        if (scheduleSetupAttempt.Success is false)
-        {
-            throw new Exception("Setup failed");
-        }
-
-        var scheduleAttempt = await ContentPublishingService.PublishAsync(
-            setupData.Key,
-            [
-                new()
-                {
-                    Culture = setupInfo.LangEn.IsoCode,
-                    Schedule = new ContentScheduleModel(),
-                },
-            ],
-            Constants.Security.SuperUserKey);
-
-        Assert.IsTrue(scheduleAttempt.Success);
-
-        var schedules = ContentService.GetContentScheduleByContentId(setupData.Id);
-        var content = ContentService.GetById(setupData.Key);
-
-        Assert.Multiple(() =>
-        {
-            Assert.AreEqual(0, content!.PublishedCultures.Count());
-            Assert.IsFalse(schedules.GetSchedule(setupInfo.LangEn.IsoCode, ContentScheduleAction.Release).Any());
-            Assert.IsFalse(schedules.GetSchedule(setupInfo.LangEn.IsoCode, ContentScheduleAction.Expire).Any());
-            Assert.AreEqual(4, schedules.FullSchedule.Count);
-        });
-    }
-
-    [Test]
-    public async Task Can_Clear_Schedule_Some_Cultures()
-    {
-        var setupInfo = await SetupVariantDoctypeAsync();
-        var setupData = await CreateVariantContentAsync(
-            setupInfo.LangEn,
-            setupInfo.LangDa,
-            setupInfo.LangBe,
-            setupInfo.contentType);
-
-        var scheduleSetupAttempt =
-            await SchedulePublishAndUnPublishForAllCulturesAsync(setupData, setupInfo);
-
-        if (scheduleSetupAttempt.Success is false)
-        {
-            throw new Exception("Setup failed");
-        }
-
-        var scheduleAttempt = await ContentPublishingService.PublishAsync(
-            setupData.Key,
-            [
-                new()
-                {
-                    Culture = setupInfo.LangEn.IsoCode,
-                    Schedule = new ContentScheduleModel(),
-                },
-                new()
-                {
-                    Culture = setupInfo.LangDa.IsoCode,
-                    Schedule = new ContentScheduleModel(),
-                },
-            ],
-            Constants.Security.SuperUserKey);
-
-        Assert.IsTrue(scheduleAttempt.Success);
-
-        var schedules = ContentService.GetContentScheduleByContentId(setupData.Id);
-        var content = ContentService.GetById(setupData.Key);
-
-        Assert.Multiple(() =>
-        {
-            Assert.AreEqual(0, content!.PublishedCultures.Count());
-            Assert.IsFalse(schedules.GetSchedule(setupInfo.LangEn.IsoCode, ContentScheduleAction.Release).Any());
-            Assert.IsFalse(schedules.GetSchedule(setupInfo.LangEn.IsoCode, ContentScheduleAction.Expire).Any());
-            Assert.IsFalse(schedules.GetSchedule(setupInfo.LangDa.IsoCode, ContentScheduleAction.Release).Any());
-            Assert.IsFalse(schedules.GetSchedule(setupInfo.LangDa.IsoCode, ContentScheduleAction.Expire).Any());
-            Assert.AreEqual(2, schedules.FullSchedule.Count);
-        });
-    }
-
-    [Test]
-    public async Task Can_Clear_Schedule_All_Cultures()
-    {
-        var setupInfo = await SetupVariantDoctypeAsync();
-        var setupData = await CreateVariantContentAsync(
-            setupInfo.LangEn,
-            setupInfo.LangDa,
-            setupInfo.LangBe,
-            setupInfo.contentType);
-
-        var scheduleSetupAttempt =
-            await SchedulePublishAndUnPublishForAllCulturesAsync(setupData, setupInfo);
-
-        if (scheduleSetupAttempt.Success is false)
-        {
-            throw new Exception("Setup failed");
-        }
-
-        var scheduleAttempt = await ContentPublishingService.PublishAsync(
-            setupData.Key,
-            [
-                new()
-                {
-                    Culture = setupInfo.LangEn.IsoCode,
-                    Schedule = new ContentScheduleModel(),
-                },
-                new()
-                {
-                    Culture = setupInfo.LangDa.IsoCode,
-                    Schedule = new ContentScheduleModel(),
-                },
-                new()
-                {
-                    Culture = setupInfo.LangBe.IsoCode,
-                    Schedule = new ContentScheduleModel(),
-                },
-            ],
-            Constants.Security.SuperUserKey);
-
-        Assert.IsTrue(scheduleAttempt.Success);
-
-        var schedules = ContentService.GetContentScheduleByContentId(setupData.Id);
-        var content = ContentService.GetById(setupData.Key);
-
-        Assert.Multiple(() =>
-        {
-            Assert.AreEqual(0, content!.PublishedCultures.Count());
-            Assert.AreEqual(0, schedules.FullSchedule.Count);
-        });
-    }
-
-    #endregion
-
-    #region Combinations
-
-    [Test]
-    public async Task Can_Publish_And_Schedule_Different_Cultures()
-    {
-        var (langEn, langDa, langBe, contentType) = await SetupVariantDoctypeAsync();
-        var setupData = await CreateVariantContentAsync(
-            langEn,
-            langDa,
-            langBe,
-            contentType);
-
-        var scheduleAttempt = await ContentPublishingService.PublishAsync(
-            setupData.Key,
-            [
-                new()
-                {
-                    Culture = langEn.IsoCode,
-                },
-                new()
-                {
-                    Culture = langDa.IsoCode,
-                    Schedule = new ContentScheduleModel { PublishDate = _schedulePublishDate },
-                },
-            ],
-            Constants.Security.SuperUserKey);
-
-        Assert.IsTrue(scheduleAttempt.Success);
-
-        var schedules = ContentService.GetContentScheduleByContentId(setupData.Id);
-        var content = ContentService.GetById(setupData.Key);
-
-        Assert.Multiple(() =>
-        {
-            Assert.AreEqual(1, content!.PublishedCultures.Count());
-            Assert.AreEqual(1, schedules.FullSchedule.Count);
-        });
-    }
-    #endregion
-
-    #region Helper methods
 
     private async Task<(ILanguage LangEn, ILanguage LangDa, ILanguage LangBe, IContentType contentType)>
         SetupVariantDoctypeAsync()
@@ -1357,10 +60,10 @@ public class ContentPublishingServiceTests : UmbracoIntegrationTestWithContent
             .WithName("Variant Content")
             .WithContentVariation(ContentVariation.Culture)
             .AddPropertyGroup()
-            .WithAlias("content")
-            .WithName("Content")
-            .WithSupportsPublishing(true)
-            .Done()
+                .WithAlias("content")
+                .WithName("Content")
+                .WithSupportsPublishing(true)
+                .Done()
             .Build();
 
         contentType.AllowedAsRoot = true;
@@ -1509,6 +212,4 @@ public class ContentPublishingServiceTests : UmbracoIntegrationTestWithContent
                 },
             ],
             Constants.Security.SuperUserKey);
-
-    #endregion
 }
